+++ cpp/tests/join/join_tests.cpp
@@ -27,6 +27,7 @@
 #include <tests/utilities/column_wrapper.hpp>
 #include <tests/utilities/table_utilities.hpp>
 #include <tests/utilities/type_lists.hpp>
+#include "cudf/types.hpp"
 
 template <typename T>
 using column_wrapper = cudf::test::fixed_width_column_wrapper<T>;
@@ -212,6 +213,100 @@ TEST_F(JoinTest, FullJoinWithNulls)
   cudf::test::expect_tables_equal(*sorted_gold, *sorted_result);
 }
 
+TEST_F(JoinTest, FullJoinOnNulls)
+{
+  // clang-format off
+  column_wrapper<int32_t> col0_0{{  3,    1 },
+                                 {  1,    0  }};
+  strcol_wrapper          col0_1({"s0", "s1" });
+  column_wrapper<int32_t> col0_2{{  0,    1 }};
+
+  column_wrapper<int32_t> col1_0{{  2,    5,    3,    7 },
+                                 {  1,    1,    1,    0 }};
+  strcol_wrapper          col1_1({"s1", "s0", "s0", "s1" });
+  column_wrapper<int32_t> col1_2{{  1,    4,    2,    8 }};
+
+  CVector cols0, cols1;
+  cols0.push_back(col0_0.release());
+  cols0.push_back(col0_1.release());
+  cols0.push_back(col0_2.release());
+  cols1.push_back(col1_0.release());
+  cols1.push_back(col1_1.release());
+  cols1.push_back(col1_2.release());
+
+  Table t0(std::move(cols0));
+  Table t1(std::move(cols1));
+
+  auto result            = cudf::full_join(t0, t1, {0, 1}, {0, 1}, {{0, 0}, {1, 1}});
+  auto result_sort_order = cudf::sorted_order(result->view());
+  auto sorted_result     = cudf::gather(result->view(), *result_sort_order);
+
+#if 0
+  std::cout << "Actual Results:\n";
+  cudf::test::print(sorted_result->get_column(0).view(), std::cout, ",\t\t");
+  cudf::test::print(sorted_result->get_column(1).view(), std::cout, ",\t\t");
+  cudf::test::print(sorted_result->get_column(2).view(), std::cout, ",\t\t");
+  cudf::test::print(sorted_result->get_column(3).view(), std::cout, ",\t\t");
+#endif
+ 
+  column_wrapper<int32_t> col_gold_0{{   2,    5,    3,    -1},
+                                     {   1,    1,    1,     0}};
+  strcol_wrapper          col_gold_1({ "s1", "s0", "s0",  "s1"});
+  column_wrapper<int32_t> col_gold_2{{  -1,   -1,    0,     1}, 
+                                     {   0,    0,    1,     1}};
+  column_wrapper<int32_t> col_gold_3{{   1,    4,    2,     8}, 
+                                     {   1,    1,    1,     1}};
+
+  CVector cols_gold;
+  cols_gold.push_back(col_gold_0.release());
+  cols_gold.push_back(col_gold_1.release());
+  cols_gold.push_back(col_gold_2.release());
+  cols_gold.push_back(col_gold_3.release());
+  Table gold(std::move(cols_gold));
+
+  auto gold_sort_order = cudf::sorted_order(gold.view());
+  auto sorted_gold     = cudf::gather(gold.view(), *gold_sort_order);
+
+#if 0
+  std::cout << "Expected Results:\n";
+  cudf::test::print(sorted_gold->get_column(0).view(), std::cout, ",\t\t");
+  cudf::test::print(sorted_gold->get_column(1).view(), std::cout, ",\t\t");
+  cudf::test::print(sorted_gold->get_column(2).view(), std::cout, ",\t\t");
+  cudf::test::print(sorted_gold->get_column(3).view(), std::cout, ",\t\t");
+#endif
+
+  cudf::test::expect_tables_equal(*sorted_gold, *sorted_result);
+
+  // Repeat test with compare_nulls_equal=false,
+  // as per SQL standard.
+
+  result            = cudf::full_join(t0, t1, {0, 1}, {0, 1}, {{0, 0}, {1, 1}}, cudf::null_equality::UNEQUAL);
+  result_sort_order = cudf::sorted_order(result->view());
+  sorted_result     = cudf::gather(result->view(), *result_sort_order);
+
+  col_gold_0 =               {{   2,    5,    3,    -1,   -1},
+                              {   1,    1,    1,     0,    0}};
+  col_gold_1 = strcol_wrapper({ "s1", "s0", "s0",  "s1", "s1"});
+  col_gold_2 =               {{  -1,   -1,    0,    -1,    1}, 
+                              {   0,    0,    1,     0,    1}};
+  col_gold_3 =               {{   1,    4,    2,     8,   -1}, 
+                              {   1,    1,    1,     1,    0}};
+
+  // clang-format on
+
+  CVector cols_gold_nulls_unequal;
+  cols_gold_nulls_unequal.push_back(col_gold_0.release());
+  cols_gold_nulls_unequal.push_back(col_gold_1.release());
+  cols_gold_nulls_unequal.push_back(col_gold_2.release());
+  cols_gold_nulls_unequal.push_back(col_gold_3.release());
+  Table gold_nulls_unequal{std::move(cols_gold_nulls_unequal)};
+
+  gold_sort_order = cudf::sorted_order(gold_nulls_unequal.view());
+  sorted_gold     = cudf::gather(gold_nulls_unequal.view(), *gold_sort_order);
+
+  cudf::test::expect_tables_equal(*sorted_gold, *sorted_result);
+}
+
 TEST_F(JoinTest, LeftJoinNoNulls)
 {
   column_wrapper<int32_t> col0_0{{3, 1, 2, 0, 3}};
@@ -294,6 +389,101 @@ TEST_F(JoinTest, LeftJoinWithNulls)
   cudf::test::expect_tables_equal(*sorted_gold, *sorted_result);
 }
 
+TEST_F(JoinTest, LeftJoinOnNulls)
+{
+  // clang-format off
+  column_wrapper<int32_t> col0_0{{  3,    1,    2},
+                                 {  1,    0,    1}};
+  strcol_wrapper          col0_1({"s0", "s1", "s2" });
+  column_wrapper<int32_t> col0_2{{  0,    1,    2 }};
+
+  column_wrapper<int32_t> col1_0{{  2,    5,    3,    7 },
+                                 {  1,    1,    1,    0 }};
+  strcol_wrapper          col1_1({"s1", "s0", "s0", "s1" });
+  column_wrapper<int32_t> col1_2{{  1,    4,    2,    8 }};
+
+  CVector cols0, cols1;
+  cols0.push_back(col0_0.release());
+  cols0.push_back(col0_1.release());
+  cols0.push_back(col0_2.release());
+  cols1.push_back(col1_0.release());
+  cols1.push_back(col1_1.release());
+  cols1.push_back(col1_2.release());
+
+  Table t0(std::move(cols0));
+  Table t1(std::move(cols1));
+
+  auto result            = cudf::left_join(t0, t1, {0, 1}, {0, 1}, {{0, 0}, {1, 1}});
+  auto result_sort_order = cudf::sorted_order(result->view());
+  auto sorted_result     = cudf::gather(result->view(), *result_sort_order);
+
+#if 0
+  std::cout << "Actual Results:\n";
+  cudf::test::print(sorted_result->get_column(0).view(), std::cout, ",\t\t");
+  cudf::test::print(sorted_result->get_column(1).view(), std::cout, ",\t\t");
+  cudf::test::print(sorted_result->get_column(2).view(), std::cout, ",\t\t");
+  cudf::test::print(sorted_result->get_column(3).view(), std::cout, ",\t\t");
+#endif
+ 
+  column_wrapper<int32_t> col_gold_0{{   3,    -1,    2},
+                                     {   1,     0,    1}};
+  strcol_wrapper          col_gold_1({ "s0",  "s1", "s2"},
+                                     {   1,     1,    1});
+  column_wrapper<int32_t> col_gold_2{{   0,     1,    2}, 
+                                     {   1,     1,    1}};
+  column_wrapper<int32_t> col_gold_3{{   2,     8,   -1}, 
+                                     {   1,     1,    0}};
+
+  CVector cols_gold;
+  cols_gold.push_back(col_gold_0.release());
+  cols_gold.push_back(col_gold_1.release());
+  cols_gold.push_back(col_gold_2.release());
+  cols_gold.push_back(col_gold_3.release());
+  Table gold(std::move(cols_gold));
+
+  auto gold_sort_order = cudf::sorted_order(gold.view());
+  auto sorted_gold     = cudf::gather(gold.view(), *gold_sort_order);
+
+#if 0
+  std::cout << "Expected Results:\n";
+  cudf::test::print(sorted_gold->get_column(0).view(), std::cout, ",\t\t");
+  cudf::test::print(sorted_gold->get_column(1).view(), std::cout, ",\t\t");
+  cudf::test::print(sorted_gold->get_column(2).view(), std::cout, ",\t\t");
+  cudf::test::print(sorted_gold->get_column(3).view(), std::cout, ",\t\t");
+#endif
+
+  cudf::test::expect_tables_equal(*sorted_gold, *sorted_result);
+
+  // Repeat test with compare_nulls_equal=false,
+  // as per SQL standard.
+
+  result            = cudf::left_join(t0, t1, {0, 1}, {0, 1}, {{0, 0}, {1, 1}}, cudf::null_equality::UNEQUAL);
+  result_sort_order = cudf::sorted_order(result->view());
+  sorted_result     = cudf::gather(result->view(), *result_sort_order);
+
+  col_gold_0 =               {{   3,    -1,    2},
+                              {   1,     0,    1}};
+  col_gold_1 = strcol_wrapper({ "s0",  "s1", "s2"},
+                              {   1,     1,    1});
+  col_gold_2 =               {{   0,     1,    2}, 
+                              {   1,     1,    1}};
+  col_gold_3 =               {{   2,    -1,   -1}, 
+                              {   1,     0,    0}};
+
+  // clang-format on
+  CVector cols_gold_nulls_unequal;
+  cols_gold_nulls_unequal.push_back(col_gold_0.release());
+  cols_gold_nulls_unequal.push_back(col_gold_1.release());
+  cols_gold_nulls_unequal.push_back(col_gold_2.release());
+  cols_gold_nulls_unequal.push_back(col_gold_3.release());
+  Table gold_nulls_unequal{std::move(cols_gold_nulls_unequal)};
+
+  gold_sort_order = cudf::sorted_order(gold_nulls_unequal.view());
+  sorted_gold     = cudf::gather(gold_nulls_unequal.view(), *gold_sort_order);
+
+  cudf::test::expect_tables_equal(*sorted_gold, *sorted_result);
+}
+
 TEST_F(JoinTest, InnerJoinNoNulls)
 {
   column_wrapper<int32_t> col0_0{{3, 1, 2, 0, 2}};
@@ -373,6 +563,78 @@ TEST_F(JoinTest, InnerJoinWithNulls)
 
   auto gold_sort_order = cudf::sorted_order(gold.view());
   auto sorted_gold     = cudf::gather(gold.view(), *gold_sort_order);
+  cudf::test::expect_tables_equal(*sorted_gold, *sorted_result);
+}
+
+// Test to check join behaviour when join keys are null.
+TEST_F(JoinTest, InnerJoinOnNulls)
+{
+  // clang-format off
+  column_wrapper<int32_t> col0_0{{  3,    1,    2,    0,    2}};
+  strcol_wrapper          col0_1({"s1", "s1", "s8", "s4", "s0"}, 
+                                 {  1,    1,    0,    1,    1});
+  column_wrapper<int32_t> col0_2{{  0,    1,    2,    4,    1}};
+
+  column_wrapper<int32_t> col1_0{{  2,    2,    0,    4,    3}};
+  strcol_wrapper          col1_1({"s1", "s0", "s1", "s2", "s1"}, 
+                                 {  1,    0,    1,    1,    1});
+  column_wrapper<int32_t> col1_2{{  1,    0,    1,    2,    1}};
+
+  CVector cols0, cols1;
+  cols0.push_back(col0_0.release());
+  cols0.push_back(col0_1.release());
+  cols0.push_back(col0_2.release());
+  cols1.push_back(col1_0.release());
+  cols1.push_back(col1_1.release());
+  cols1.push_back(col1_2.release());
+
+  Table t0(std::move(cols0));
+  Table t1(std::move(cols1));
+
+  auto result            = cudf::inner_join(t0, t1, {0, 1}, {0, 1}, {{0, 0}, {1, 1}});
+  auto result_sort_order = cudf::sorted_order(result->view());
+  auto sorted_result     = cudf::gather(result->view(), *result_sort_order);
+
+  column_wrapper<int32_t> col_gold_0 {{  3,    2}};
+  strcol_wrapper          col_gold_1 ({"s1", "s0"}, 
+                                      {  1,    0});
+  column_wrapper<int32_t> col_gold_2{{   0,    2}};
+  column_wrapper<int32_t> col_gold_3{{   1,    0}};
+  CVector cols_gold;
+  cols_gold.push_back(col_gold_0.release());
+  cols_gold.push_back(col_gold_1.release());
+  cols_gold.push_back(col_gold_2.release());
+  cols_gold.push_back(col_gold_3.release());
+  Table gold(std::move(cols_gold));
+
+  auto gold_sort_order = cudf::sorted_order(gold.view());
+  auto sorted_gold     = cudf::gather(gold.view(), *gold_sort_order);
+  cudf::test::expect_tables_equal(*sorted_gold, *sorted_result);
+  
+  // Repeat test with compare_nulls_equal=false,
+  // as per SQL standard.
+
+  result            = cudf::inner_join(t0, t1, {0, 1}, {0, 1}, {{0, 0}, {1, 1}}, cudf::null_equality::UNEQUAL);
+  result_sort_order = cudf::sorted_order(result->view());
+  sorted_result     = cudf::gather(result->view(), *result_sort_order);
+
+  col_gold_0 =               {{  3}};
+  col_gold_1 = strcol_wrapper({"s1"}, 
+                              {  1});
+  col_gold_2 =               {{  0}};
+  col_gold_3 =               {{  1}};
+
+  // clang-format on
+
+  CVector cols_gold_sql;
+  cols_gold_sql.push_back(col_gold_0.release());
+  cols_gold_sql.push_back(col_gold_1.release());
+  cols_gold_sql.push_back(col_gold_2.release());
+  cols_gold_sql.push_back(col_gold_3.release());
+  Table gold_sql(std::move(cols_gold_sql));
+
+  gold_sort_order = cudf::sorted_order(gold_sql.view());
+  sorted_gold     = cudf::gather(gold_sql.view(), *gold_sort_order);
   cudf::test::expect_tables_equal(*sorted_gold, *sorted_result);
 }