+++ python/cudf/cudf/comm/gpuarrow.py
@@ -1,5 +1,4 @@
 # Copyright (c) 2019, NVIDIA CORPORATION.
-
 from collections import OrderedDict
 from collections.abc import Sequence